Clear warnings from GCC.

+++ src/protocols/ab/eip_plc5_dhp.c
@@ -194,7 +194,7 @@ int tag_tickler(ab_tag_p tag)
             /* read done so create done. */
             if(tag->first_read) {
                 tag->first_read = 0;
-                tag_raise_event((plc_tag_p)tag, PLCTAG_EVENT_CREATED, rc);
+                tag_raise_event((plc_tag_p)tag, PLCTAG_EVENT_CREATED, (int8_t)rc);
             }
 
             tag->read_complete = 1;
